Changed Sprite to inherit from GameObject
Sprite:
-Sprite now has a constructor that can set the default sequence it will
show when rendered.  If no sequence is giver, then the Sprite will show
the first frame if it has one.
-Added update().  update currently nothing in the Sprite class and is
meant for other classes that inherit from the Sprite class to implement.
-Added handleEvent().  handleEvent currently does nothing and is meant
to be implemented by classes that inherit from the Sprite class.
-Added render().  render is what is called to show the Sprite on the
screen.
-Removed currX, currY, width, height variables.  This information is now
stored in the Body struct that is inherited from GameObject.