Added Character class
Character:
-Character inherits from the Sprite class and has one more variable that
can be modified in the constructor.  The health of the character is 100
by default.
-getHP() gets the current HP of the character.  changeHP() adds or
removes health from the character.
-Movement for the character is done by the functions moveLeft(),
moveRight(), jump(), fall().  The struct moveSeq stores the sequences
that these functions will use to change how the character is shown on
the screen during one of these actions.  By default the sequences are
set to "walk left", "walk right", "walk up", and "walk down"
respectively.
-changeSequenceName() changes the above default sequence names to
whatever the user wants them to be.  The first parameter is the action
type which is either "left", "right", "jump", "fall" and the second
parameter is what you want the sequence name to be.
-render() shows the current sequence that the character is in.
-The character has a mass variable that doesn't do anything until the
Physics class is added to the character.
TODO: Fix update() and handleEvent() functions to take in events.